BAH-4888 | Refactor. Remove Translator type, move formatSearchResult to widgets, and inline translate fn signatures
