+++ packages/bahmni-services/src/index.ts
@@ -113,10 +113,8 @@ export {
   camelToScreamingSnakeCase,
   convertToSentenceCase,
   resolveComboBoxItems,
-  type Translator,
   formatGender,
   formatCountry,
-  formatSearchResult,
 } from './utils';
 export {
   type FormatDateResult,
