+++ packages/bahmni-services/src/utils/index.ts
@@ -17,8 +17,6 @@ export {
   camelToScreamingSnakeCase,
   convertToSentenceCase,
   resolveComboBoxItems,
-  type Translator,
   formatGender,
   formatCountry,
-  formatSearchResult,
 } from './utils';
